requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
openai
elevenlabs
azure-cognitiveservices-speech
silero-vad
torch
